% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{asn_attachments}
\alias{asn_attachments}
\title{Attachments}
\description{
An _attachment_ object represents any file attached to a task in Asana,
 whether it's an uploaded file or one associated via a third-party service
 such as Dropbox or Google Drive.
}
\seealso{
\link{https://asana.com/developers/api-reference/attachments}

Other attachments: \code{\link{asn_attachments_find_by_id}},
  \code{\link{asn_attachments_find_by_task}}
}
