% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_field_settings.R
\name{asn_custom_field_settings}
\alias{asn_custom_field_settings}
\title{CustomFieldSettings}
\description{
Custom fields are attached to a particular project with the Custom
 Field Settings resource. This resource both represents the many-to-many join
 of the Custom Field and Project as well as stores information that is relevant to that
 particular pairing; for instance, the `is_important` property determines
 some possible application-specific handling of that custom field (see below)
}
\seealso{
\link{https://asana.com/developers/api-reference/custom_field_settings}

Other custom_field_settings: \code{\link{asn_custom_field_settings_find_by_project}}
}
\concept{custom_field_settings}
