% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fields.R
\name{asn_custom_fields}
\alias{asn_custom_fields}
\title{CustomFields}
\description{
Custom Fields store the metadata that is used in order to add user-specified
 information to tasks in Asana. Be sure to reference the [Custom
 Fields](/developers/documentation/getting-started/custom-fields) developer
 documentation for more information about how custom fields relate to various
 resources in Asana.
}
\seealso{
\link{https://asana.com/developers/api-reference/custom_fields}

Other custom_fields: \code{\link{asn_custom_fields_add_enum_option}},
  \code{\link{asn_custom_fields_create}},
  \code{\link{asn_custom_fields_delete}},
  \code{\link{asn_custom_fields_find_by_id}},
  \code{\link{asn_custom_fields_find_by_workspace}},
  \code{\link{asn_custom_fields_reorder_enum_option}},
  \code{\link{asn_custom_fields_update_enum_option}},
  \code{\link{asn_custom_fields_update}}
}
