% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fields.R
\name{asn_custom_fields_add_enum_option}
\alias{asn_custom_fields_add_enum_option}
\title{Creates an enum option and adds it to this custom field's list of enum options. A custom field can have at most 50 enum options (including disabled options). By default new enum options are inserted at the end of a custom field's list.}
\usage{
asn_custom_fields_add_enum_option(custom_field, ..., options = list())
}
\arguments{
\item{custom_field}{{Id} Globally unique identifier for the custom field.#'}

\item{...}{{Object} Data for the request
name : {String} The name of the enum option.
[color] : {String} The color of the enum option. Defaults to 'none'.
[insert_before] : {Id} An existing enum option within this custom field before which the new enum option should be inserted. Cannot be provided together with after_enum_option.
[insert_after] : {Id} An existing enum option within this custom field after which the new enum option should be inserted. Cannot be provided together with before_enum_option.}
}
\description{
Returns the full record of the newly created enum option.
}
\seealso{
Other custom_fields: \code{\link{asn_custom_fields_create}},
  \code{\link{asn_custom_fields_delete}},
  \code{\link{asn_custom_fields_find_by_id}},
  \code{\link{asn_custom_fields_find_by_workspace}},
  \code{\link{asn_custom_fields_reorder_enum_option}},
  \code{\link{asn_custom_fields_update_enum_option}},
  \code{\link{asn_custom_fields_update}},
  \code{\link{asn_custom_fields}}
}
\concept{custom_fields}
