% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fields.R
\name{asn_custom_fields_create}
\alias{asn_custom_fields_create}
\title{Creates a new custom field in a workspace. Every custom field is required to be created in a specific workspace, and this workspace cannot be changed once set.}
\usage{
asn_custom_fields_create(..., options = list())
}
\arguments{
\item{...}{{Object} Data for the request
workspace : {Id} The workspace to create a custom field in.
type : {String} The type of the custom field.
name : {String} The name of the custom field.
[description] : {String} The description of the custom field.
[precision] : {Integer} The number of decimal places for the numerical values. Required if the custom field is of type 'number'.
[enum_options] : {String} The discrete values the custom field can assume. Required if the custom field is of type 'enum'.}
}
\description{
A custom field's `name` must be unique within a workspace and not conflict with names of existing task properties such as 'Due Date' or 'Assignee'. A custom field's `type` must be one of  'text', 'enum', or 'number'.
}
\details{
Returns the full record of the newly created custom field.
}
\seealso{
Other custom_fields: \code{\link{asn_custom_fields_add_enum_option}},
  \code{\link{asn_custom_fields_delete}},
  \code{\link{asn_custom_fields_find_by_id}},
  \code{\link{asn_custom_fields_find_by_workspace}},
  \code{\link{asn_custom_fields_reorder_enum_option}},
  \code{\link{asn_custom_fields_update_enum_option}},
  \code{\link{asn_custom_fields_update}},
  \code{\link{asn_custom_fields}}
}
