% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fields.R
\name{asn_custom_fields_delete}
\alias{asn_custom_fields_delete}
\title{A specific, existing custom field can be deleted by making a DELETE request on the URL for that custom field.}
\usage{
asn_custom_fields_delete(custom_field, ..., options = list())
}
\arguments{
\item{custom_field}{{Id} Globally unique identifier for the custom field.}
}
\description{
Returns an empty data record.
}
