% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fields.R
\name{asn_custom_fields_reorder_enum_option}
\alias{asn_custom_fields_reorder_enum_option}
\title{Moves a particular enum option to be either before or after another specified enum option in the custom field.}
\usage{
asn_custom_fields_reorder_enum_option(custom_field, ...,
  options = list())
}
\arguments{
\item{custom_field}{{Id} Globally unique identifier for the custom field.#'}

\item{...}{{Object} Data for the request
enum_option : {Id} The ID of the enum option to relocate.
name : {String} The name of the enum option.
[color] : {String} The color of the enum option. Defaults to 'none'.
[before_enum_option] : {Id} An existing enum option within this custom field before which the new enum option should be inserted. Cannot be provided together with after_enum_option.
[after_enum_option] : {Id} An existing enum option within this custom field after which the new enum option should be inserted. Cannot be provided together with before_enum_option.}
}
\description{
Moves a particular enum option to be either before or after another specified enum option in the custom field.
}
\seealso{
Other custom_fields: \code{\link{asn_custom_fields_add_enum_option}},
  \code{\link{asn_custom_fields_create}},
  \code{\link{asn_custom_fields_delete}},
  \code{\link{asn_custom_fields_find_by_id}},
  \code{\link{asn_custom_fields_find_by_workspace}},
  \code{\link{asn_custom_fields_update_enum_option}},
  \code{\link{asn_custom_fields_update}},
  \code{\link{asn_custom_fields}}
}
\concept{custom_fields}
