% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fields.R
\name{asn_custom_fields_update_enum_option}
\alias{asn_custom_fields_update_enum_option}
\title{Updates an existing enum option. Enum custom fields require at least one enabled enum option.}
\usage{
asn_custom_fields_update_enum_option(enum_option, ..., options = list())
}
\arguments{
\item{enum_option}{{Id} Globally unique identifier for the enum option.#'}

\item{...}{{Object} Data for the request
name : {String} The name of the enum option.
[color] : {String} The color of the enum option. Defaults to 'none'.
[enabled] : {Boolean} Whether or not the enum option is a selectable value for the custom field.}
}
\description{
Returns the full record of the updated enum option.
}
\seealso{
Other custom_fields: \code{\link{asn_custom_fields_add_enum_option}},
  \code{\link{asn_custom_fields_create}},
  \code{\link{asn_custom_fields_delete}},
  \code{\link{asn_custom_fields_find_by_id}},
  \code{\link{asn_custom_fields_find_by_workspace}},
  \code{\link{asn_custom_fields_reorder_enum_option}},
  \code{\link{asn_custom_fields_update}},
  \code{\link{asn_custom_fields}}
}
\concept{custom_fields}
