% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization_exports.R
\name{asn_organization_exports}
\alias{asn_organization_exports}
\title{OrganizationExports}
\description{
An _organization_export_ object represents a request to export the complete data of an Organization
 in JSON format.
 
 To export an Organization using this API:
 
 * Create an `organization_export` [request](#create) and store the id that is returned.\
 * Request the `organization_export` every few minutes, until the `state` field contains 'finished'.\
 * Download the file located at the URL in the `download_url` field.
 
 Exports can take a long time, from several minutes to a few hours for large Organizations.
 
 **Note:** These endpoints are only available to [Service Accounts](/guide/help/premium/service-accounts)
 of an [Enterprise](/enterprise) Organization.
}
\seealso{
\link{https://asana.com/developers/api-reference/organization_exports}

Other organization_exports: \code{\link{asn_organization_exports_create}},
  \code{\link{asn_organization_exports_find_by_id}}
}
