% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization_exports.R
\name{asn_organization_exports_create}
\alias{asn_organization_exports_create}
\title{This method creates a request to export an Organization. Asana will complete the export at some
point after you create the request.}
\usage{
asn_organization_exports_create(..., options = list())
}
\arguments{
\item{...}{{Object} Data for the request
organization : {Id} Globally unique identifier for the workspace or organization.}
}
\description{
This method creates a request to export an Organization. Asana will complete the export at some
point after you create the request.
}
\seealso{
Other organization_exports: \code{\link{asn_organization_exports_find_by_id}},
  \code{\link{asn_organization_exports}}
}
\concept{organization_exports}
