% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{asn_post}
\alias{asn_post}
\title{Dispatch a POST request to Asana}
\usage{
asn_post(endpoint, ..., options = list(),
  .token = Sys.getenv("ASANA_ACCESS_TOKEN"))
}
\arguments{
\item{endpoint}{endpoint}

\item{...}{query parameters}

\item{.token}{access token}
}
\description{
Dispatch a POST request to Asana
}
\examples{
\dontrun{
  asn_post('/teams/730118080732127/projects', name = 'My New Project')
}
}
