% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_memberships.R
\name{asn_project_memberships}
\alias{asn_project_memberships}
\title{ProjectMemberships}
\description{
With the introduction of "comment-only" projects in Asana, a user's membership
 in a project comes with associated permissions. These permissions (whether a
 user has full access to the project or comment-only access) are accessible
 through the project memberships endpoints described here.
}
\seealso{
\link{https://asana.com/developers/api-reference/project_memberships}

Other project_memberships: \code{\link{asn_project_memberships_get_many}},
  \code{\link{asn_project_memberships_get_single}}
}
\concept{project_memberships}
