% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_memberships.R
\name{asn_project_memberships_get_many}
\alias{asn_project_memberships_get_many}
\title{Returns the compact project membership records for the project.}
\usage{
asn_project_memberships_get_many(project, ..., options = list())
}
\arguments{
\item{project}{{Id} The project for which to fetch memberships.#'}

\item{...}{{Object} Parameters for the request
[user] : {String} If present, the user to filter the memberships to.}
}
\description{
Returns the compact project membership records for the project.
}
