% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_memberships.R
\name{asn_project_memberships_get_single}
\alias{asn_project_memberships_get_single}
\title{Returns the project membership record.}
\usage{
asn_project_memberships_get_single(project, ..., options = list())
}
\arguments{
\item{project}{{Id} Globally unique identifier for the project membership.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
Returns the project membership record.
}
\seealso{
Other project_memberships: \code{\link{asn_project_memberships_get_many}},
  \code{\link{asn_project_memberships}}
}
