% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_statuses.R
\name{asn_project_statuses}
\alias{asn_project_statuses}
\title{ProjectStatuses}
\description{
A _project status_ is an update on the progress of a particular project, and is sent out to all project
 followers when created. These updates include both text describing the update and a color code intended to
 represent the overall state of the project: "green" for projects that are on track, "yellow" for projects
 at risk, and "red" for projects that are behind.
 
 Project statuses can be created and deleted, but not modified.
}
\seealso{
\link{https://asana.com/developers/api-reference/project_statuses}

Other project_statuses: \code{\link{asn_project_statuses_create}},
  \code{\link{asn_project_statuses_delete}},
  \code{\link{asn_project_statuses_find_by_id}},
  \code{\link{asn_project_statuses_find_by_project}}
}
