% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_statuses.R
\name{asn_project_statuses_create}
\alias{asn_project_statuses_create}
\title{Creates a new status update on the project.}
\usage{
asn_project_statuses_create(project, ..., options = list())
}
\arguments{
\item{project}{{Id} The project on which to create a status update.#'}

\item{...}{{Object} Data for the request
text : {String} The text of the project status update.
color : {String} The color to associate with the status update. Must be one of `"red"`, `"yellow"`, or `"green"`.}
}
\description{
Returns the full record of the newly created project status update.
}
\seealso{
Other project_statuses: \code{\link{asn_project_statuses_delete}},
  \code{\link{asn_project_statuses_find_by_id}},
  \code{\link{asn_project_statuses_find_by_project}},
  \code{\link{asn_project_statuses}}
}
