% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_statuses.R
\name{asn_project_statuses_delete}
\alias{asn_project_statuses_delete}
\title{Deletes a specific, existing project status update.}
\usage{
asn_project_statuses_delete(project_status, ..., options = list())
}
\arguments{
\item{project-status}{{Id} The project status update to delete.}
}
\description{
Returns an empty data record.
}
\seealso{
Other project_statuses: \code{\link{asn_project_statuses_create}},
  \code{\link{asn_project_statuses_find_by_id}},
  \code{\link{asn_project_statuses_find_by_project}},
  \code{\link{asn_project_statuses}}
}
