% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_statuses.R
\name{asn_project_statuses_find_by_project}
\alias{asn_project_statuses_find_by_project}
\title{Returns the compact project status update records for all updates on the project.}
\usage{
asn_project_statuses_find_by_project(project, ..., options = list())
}
\arguments{
\item{project}{{Id} The project to find status updates for.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
Returns the compact project status update records for all updates on the project.
}
