% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_projects}
\alias{asn_projects}
\title{Projects}
\description{
A _project_ represents a prioritized list of tasks in Asana or a board with
 columns of tasks represented as cards. It exists in a single workspace or
 organization and is accessible to a subset of users in that workspace or
 organization, depending on its permissions.
 
 Projects in organizations are shared with a single team. You cannot currently
 change the team of a project via the API. Non-organization workspaces do not
 have teams and so you should not specify the team of project in a regular
 workspace.
}
\seealso{
\link{https://asana.com/developers/api-reference/projects}

Other projects: \code{\link{asn_projects_add_custom_field_setting}},
  \code{\link{asn_projects_add_followers}},
  \code{\link{asn_projects_add_members}},
  \code{\link{asn_projects_create_in_team}},
  \code{\link{asn_projects_create_in_workspace}},
  \code{\link{asn_projects_create}},
  \code{\link{asn_projects_delete}},
  \code{\link{asn_projects_find_all}},
  \code{\link{asn_projects_find_by_id}},
  \code{\link{asn_projects_find_by_team}},
  \code{\link{asn_projects_find_by_workspace}},
  \code{\link{asn_projects_remove_custom_field_setting}},
  \code{\link{asn_projects_remove_followers}},
  \code{\link{asn_projects_remove_members}},
  \code{\link{asn_projects_tasks}},
  \code{\link{asn_projects_update}}
}
