% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_projects_add_custom_field_setting}
\alias{asn_projects_add_custom_field_setting}
\title{Create a new custom field setting on the project.}
\usage{
asn_projects_add_custom_field_setting(project, ..., options = list())
}
\arguments{
\item{project}{{Id} The project to associate the custom field with#'}

\item{...}{{Object} Data for the request
custom_field : {Id} The id of the custom field to associate with this project.
[is_important] : {Boolean} Whether this field should be considered important to this project.
[insert_before] : {Id} An id of a Custom Field Settings on this project, before which the new Custom Field Settings will be added.
`insert_before` and `insert_after` parameters cannot both be specified.
[insert_after] : {Id} An id of a Custom Field Settings on this project, after which the new Custom Field Settings will be added.
`insert_before` and `insert_after` parameters cannot both be specified.}
}
\description{
Create a new custom field setting on the project.
}
