% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_projects_add_followers}
\alias{asn_projects_add_followers}
\title{Adds the specified list of users as followers to the project. Followers are a subset of members, therefore if
the users are not already members of the project they will also become members as a result of this operation.
Returns the updated project record.}
\usage{
asn_projects_add_followers(project, ..., options = list())
}
\arguments{
\item{project}{{Id} The project to add followers to.#'}

\item{...}{{Object} Data for the request
followers : {Array} An array of followers to add to the project.}
}
\description{
Adds the specified list of users as followers to the project. Followers are a subset of members, therefore if
the users are not already members of the project they will also become members as a result of this operation.
Returns the updated project record.
}
