% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_projects_add_members}
\alias{asn_projects_add_members}
\title{Adds the specified list of users as members of the project. Returns the updated project record.}
\usage{
asn_projects_add_members(project, ..., options = list())
}
\arguments{
\item{project}{{Id} The project to add members to.#'}

\item{...}{{Object} Data for the request
members : {Array} An array of members to add to the project.}
}
\description{
Adds the specified list of users as members of the project. Returns the updated project record.
}
