% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_projects_create}
\alias{asn_projects_create}
\title{Creates a new project in a workspace or team.}
\usage{
asn_projects_create(..., options = list())
}
\arguments{
\item{...}{{Object} Data for the request
workspace : {Id} The workspace or organization to create the project in.
[team] : {Id} If creating in an organization, the specific team to create the
project in.}
}
\description{
Every project is required to be created in a specific workspace or
organization, and this cannot be changed once set. Note that you can use
the `workspace` parameter regardless of whether or not it is an
organization.
}
\details{
If the workspace for your project _is_ an organization, you must also
supply a `team` to share the project with.

Returns the full record of the newly created project.
}
