% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_projects_create_in_team}
\alias{asn_projects_create_in_team}
\title{Creates a project shared with the given team.}
\usage{
asn_projects_create_in_team(team, ..., options = list())
}
\arguments{
\item{team}{{Id} The team to create the project in.#'}

\item{...}{{Object} Data for the request}
}
\description{
Returns the full record of the newly created project.
}
