% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_projects_find_all}
\alias{asn_projects_find_all}
\title{Returns the compact project records for some filtered set of projects.
Use one or more of the parameters provided to filter the projects returned.}
\usage{
asn_projects_find_all(..., options = list())
}
\arguments{
\item{...}{{Object} Parameters for the request
[workspace] : {Id} The workspace or organization to filter projects on.
[team] : {Id} The team to filter projects on.
[archived] : {Boolean} Only return projects whose `archived` field takes on the value of
this parameter.}
}
\description{
Returns the compact project records for some filtered set of projects.
Use one or more of the parameters provided to filter the projects returned.
}
