% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_projects_find_by_team}
\alias{asn_projects_find_by_team}
\title{Returns the compact project records for all projects in the team.}
\usage{
asn_projects_find_by_team(team, ..., options = list())
}
\arguments{
\item{team}{{Id} The team to find projects in.#'}

\item{...}{{Object} Parameters for the request
[archived] : {Boolean} Only return projects whose `archived` field takes on the value of
this parameter.}
}
\description{
Returns the compact project records for all projects in the team.
}
