% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_projects_find_by_workspace}
\alias{asn_projects_find_by_workspace}
\title{Returns the compact project records for all projects in the workspace.}
\usage{
asn_projects_find_by_workspace(workspace, ..., options = list())
}
\arguments{
\item{workspace}{{Id} The workspace or organization to find projects in.#'}

\item{...}{{Object} Parameters for the request
[archived] : {Boolean} Only return projects whose `archived` field takes on the value of
this parameter.}
}
\description{
Returns the compact project records for all projects in the workspace.
}
