% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_projects_remove_followers}
\alias{asn_projects_remove_followers}
\title{Removes the specified list of users from following the project, this will not affect project membership status.
Returns the updated project record.}
\usage{
asn_projects_remove_followers(project, ..., options = list())
}
\arguments{
\item{project}{{Id} The project to remove followers from.#'}

\item{...}{{Object} Data for the request
followers : {Array} An array of followers to remove from the project.}
}
\description{
Removes the specified list of users from following the project, this will not affect project membership status.
Returns the updated project record.
}
