% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_projects_remove_members}
\alias{asn_projects_remove_members}
\title{Removes the specified list of members from the project. Returns the updated project record.}
\usage{
asn_projects_remove_members(project, ..., options = list())
}
\arguments{
\item{project}{{Id} The project to remove members from.#'}

\item{...}{{Object} Data for the request
members : {Array} An array of members to remove from the project.}
}
\description{
Removes the specified list of members from the project. Returns the updated project record.
}
