% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_projects_update}
\alias{asn_projects_update}
\title{A specific, existing project can be updated by making a PUT request on the
URL for that project. Only the fields provided in the `data` block will be
updated; any unspecified fields will remain unchanged.}
\usage{
asn_projects_update(project, ..., options = list())
}
\arguments{
\item{project}{{Id} The project to update.#'}

\item{...}{{Object} Data for the request}
}
\description{
When using this method, it is best to specify only those fields you wish
to change, or else you may overwrite changes made by another user since
you last retrieved the task.
}
\details{
Returns the complete updated project record.
}
\seealso{
Other projects: \code{\link{asn_projects_add_custom_field_setting}},
  \code{\link{asn_projects_add_followers}},
  \code{\link{asn_projects_add_members}},
  \code{\link{asn_projects_create_in_team}},
  \code{\link{asn_projects_create_in_workspace}},
  \code{\link{asn_projects_create}},
  \code{\link{asn_projects_delete}},
  \code{\link{asn_projects_find_all}},
  \code{\link{asn_projects_find_by_id}},
  \code{\link{asn_projects_find_by_team}},
  \code{\link{asn_projects_find_by_workspace}},
  \code{\link{asn_projects_remove_custom_field_setting}},
  \code{\link{asn_projects_remove_followers}},
  \code{\link{asn_projects_remove_members}},
  \code{\link{asn_projects_tasks}},
  \code{\link{asn_projects}}
}
\concept{projects}
