% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{asn_sections}
\alias{asn_sections}
\title{Sections}
\description{
A _section_ is a subdivision of a project that groups tasks together. It can
 either be a header above a list of tasks in a list view or a column in a
 board view of a project.
}
\seealso{
\link{https://asana.com/developers/api-reference/sections}

Other sections: \code{\link{asn_sections_create_in_project}},
  \code{\link{asn_sections_delete}},
  \code{\link{asn_sections_find_by_id}},
  \code{\link{asn_sections_find_by_project}},
  \code{\link{asn_sections_insert_in_project}},
  \code{\link{asn_sections_update}}
}
