% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{asn_sections_delete}
\alias{asn_sections_delete}
\title{A specific, existing section can be deleted by making a DELETE request
on the URL for that section.}
\usage{
asn_sections_delete(section, ..., options = list())
}
\arguments{
\item{section}{{Id} The section to delete.}
}
\description{
Note that sections must be empty to be deleted.
}
\details{
The last remaining section in a board view cannot be deleted.

Returns an empty data block.
}
