% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{asn_sections_find_by_project}
\alias{asn_sections_find_by_project}
\title{Returns the compact records for all sections in the specified project.}
\usage{
asn_sections_find_by_project(project, ..., options = list())
}
\arguments{
\item{project}{{Id} The project to get sections from.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
Returns the compact records for all sections in the specified project.
}
