% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{asn_sections_insert_in_project}
\alias{asn_sections_insert_in_project}
\title{Move sections relative to each other in a board view. One of
`before_section` or `after_section` is required.}
\usage{
asn_sections_insert_in_project(project, ..., options = list())
}
\arguments{
\item{project}{{Id} The project in which to reorder the given section#'}

\item{...}{{Object} Data for the request
section : {Id} The section to reorder
[before_section] : {Id} Insert the given section immediately before the section specified by this parameter.
[after_section] : {Id} Insert the given section immediately after the section specified by this parameter.}
}
\description{
Sections cannot be moved between projects.
}
\details{
At this point in time, moving sections is not supported in list views, only board views.

Returns an empty data block.
}
\seealso{
Other sections: \code{\link{asn_sections_create_in_project}},
  \code{\link{asn_sections_delete}},
  \code{\link{asn_sections_find_by_id}},
  \code{\link{asn_sections_find_by_project}},
  \code{\link{asn_sections_update}},
  \code{\link{asn_sections}}
}
\concept{sections}
