% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{asn_sections_update}
\alias{asn_sections_update}
\title{A specific, existing section can be updated by making a PUT request on
the URL for that project. Only the fields provided in the `data` block
will be updated; any unspecified fields will remain unchanged. (note that
at this time, the only field that can be updated is the `name` field.)}
\usage{
asn_sections_update(section, ..., options = list())
}
\arguments{
\item{section}{{Id} The section to update.#'}

\item{...}{{Object} Data for the request}
}
\description{
When using this method, it is best to specify only those fields you wish
to change, or else you may overwrite changes made by another user since
you last retrieved the task.
}
\details{
Returns the complete updated section record.
}
