% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{asn_stories}
\alias{asn_stories}
\title{Stories}
\description{
A _story_ represents an activity associated with an object in the Asana
 system. Stories are generated by the system whenever users take actions such
 as creating or assigning tasks, or moving tasks between projects. _Comments_
 are also a form of user-generated story.
 
 Stories are a form of history in the system, and as such they are read-only.
 Once generated, it is not possible to modify a story.
}
\seealso{
\link{https://asana.com/developers/api-reference/stories}

Other stories: \code{\link{asn_stories_create_on_task}},
  \code{\link{asn_stories_delete}},
  \code{\link{asn_stories_find_by_id}},
  \code{\link{asn_stories_find_by_task}},
  \code{\link{asn_stories_update}}
}
\concept{stories}
