% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{asn_stories_create_on_task}
\alias{asn_stories_create_on_task}
\title{Adds a comment to a task. The comment will be authored by the
currently authenticated user, and timestamped when the server receives
the request.}
\usage{
asn_stories_create_on_task(task, ..., options = list())
}
\arguments{
\item{task}{{Id} Globally unique identifier for the task.#'}

\item{...}{{Object} Data for the request
text : {String} The plain text of the comment to add.}
}
\description{
Returns the full record for the new story added to the task.
}
\seealso{
Other stories: \code{\link{asn_stories_delete}},
  \code{\link{asn_stories_find_by_id}},
  \code{\link{asn_stories_find_by_task}},
  \code{\link{asn_stories_update}},
  \code{\link{asn_stories}}
}
