% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{asn_stories_find_by_id}
\alias{asn_stories_find_by_id}
\title{Returns the full record for a single story.}
\usage{
asn_stories_find_by_id(story, ..., options = list())
}
\arguments{
\item{story}{{Id} Globally unique identifier for the story.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
Returns the full record for a single story.
}
\seealso{
Other stories: \code{\link{asn_stories_create_on_task}},
  \code{\link{asn_stories_delete}},
  \code{\link{asn_stories_find_by_task}},
  \code{\link{asn_stories_update}},
  \code{\link{asn_stories}}
}
\concept{stories}
