% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{asn_stories_update}
\alias{asn_stories_update}
\title{Updates the story and returns the full record for the updated story.
Only comment stories can have their text updated, and only comment stories and
attachment stories can be pinned. Only one of `text` and `html_text` can be specified.}
\usage{
asn_stories_update(story, ..., options = list())
}
\arguments{
\item{story}{{Id} Globally unique identifier for the story.#'}

\item{...}{{Object} Data for the request
[text] : {String} The plain text with which to update the comment.
[html_text] : {String} The rich text with which to update the comment.
[is_pinned] : {Boolean} Whether the story should be pinned on the resource.}
}
\description{
Updates the story and returns the full record for the updated story.
Only comment stories can have their text updated, and only comment stories and
attachment stories can be pinned. Only one of `text` and `html_text` can be specified.
}
\seealso{
Other stories: \code{\link{asn_stories_create_on_task}},
  \code{\link{asn_stories_delete}},
  \code{\link{asn_stories_find_by_id}},
  \code{\link{asn_stories_find_by_task}},
  \code{\link{asn_stories}}
}
