% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{asn_tags_create}
\alias{asn_tags_create}
\title{Creates a new tag in a workspace or organization.}
\usage{
asn_tags_create(..., options = list())
}
\arguments{
\item{...}{{Object} Data for the request
workspace : {Id} The workspace or organization to create the tag in.}
}
\description{
Every tag is required to be created in a specific workspace or
organization, and this cannot be changed once set. Note that you can use
the `workspace` parameter regardless of whether or not it is an
organization.
}
\details{
Returns the full record of the newly created tag.
}
\seealso{
Other tags: \code{\link{asn_tags_create_in_workspace}},
  \code{\link{asn_tags_delete}},
  \code{\link{asn_tags_find_all}},
  \code{\link{asn_tags_find_by_id}},
  \code{\link{asn_tags_find_by_workspace}},
  \code{\link{asn_tags_get_tasks_with_tag}},
  \code{\link{asn_tags_update}}, \code{\link{asn_tags}}
}
