% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{asn_tags_find_by_id}
\alias{asn_tags_find_by_id}
\title{Returns the complete tag record for a single tag.}
\usage{
asn_tags_find_by_id(tag, ..., options = list())
}
\arguments{
\item{tag}{{Id} The tag to get.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
Returns the complete tag record for a single tag.
}
