% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{asn_tags_update}
\alias{asn_tags_update}
\title{Updates the properties of a tag. Only the fields provided in the `data`
block will be updated; any unspecified fields will remain unchanged.}
\usage{
asn_tags_update(tag, ..., options = list())
}
\arguments{
\item{tag}{{Id} The tag to update.#'}

\item{...}{{Object} Data for the request}
}
\description{
When using this method, it is best to specify only those fields you wish
to change, or else you may overwrite changes made by another user since
you last retrieved the task.
}
\details{
Returns the complete updated tag record.
}
\seealso{
Other tags: \code{\link{asn_tags_create_in_workspace}},
  \code{\link{asn_tags_create}},
  \code{\link{asn_tags_delete}},
  \code{\link{asn_tags_find_all}},
  \code{\link{asn_tags_find_by_id}},
  \code{\link{asn_tags_find_by_workspace}},
  \code{\link{asn_tags_get_tasks_with_tag}},
  \code{\link{asn_tags}}
}
\concept{tags}
