% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_add_comment}
\alias{asn_tasks_add_comment}
\title{Adds a comment to a task. The comment will be authored by the
currently authenticated user, and timestamped when the server receives
the request.}
\usage{
asn_tasks_add_comment(task, ..., options = list())
}
\arguments{
\item{task}{{Id} Globally unique identifier for the task.#'}

\item{...}{{Object} Data for the request
text : {String} The plain text of the comment to add.}
}
\description{
Returns the full record for the new story added to the task.
}
