% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_add_dependencies}
\alias{asn_tasks_add_dependencies}
\title{Marks a set of tasks as dependencies of this task, if they are not
already dependencies. *A task can have at most 15 dependencies.*}
\usage{
asn_tasks_add_dependencies(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task to add dependencies to.#'}

\item{...}{{Object} Data for the request
dependencies : {Array} An array of task IDs that this task should depend on.}
}
\description{
Marks a set of tasks as dependencies of this task, if they are not
already dependencies. *A task can have at most 15 dependencies.*
}
