% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_add_dependents}
\alias{asn_tasks_add_dependents}
\title{Marks a set of tasks as dependents of this task, if they are not already
dependents. *A task can have at most 30 dependents.*}
\usage{
asn_tasks_add_dependents(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task to add dependents to.#'}

\item{...}{{Object} Data for the request
dependents : {Array} An array of task IDs that should depend on this task.}
}
\description{
Marks a set of tasks as dependents of this task, if they are not already
dependents. *A task can have at most 30 dependents.*
}
