% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_add_followers}
\alias{asn_tasks_add_followers}
\title{Adds each of the specified followers to the task, if they are not already
following. Returns the complete, updated record for the affected task.}
\usage{
asn_tasks_add_followers(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task to add followers to.#'}

\item{...}{{Object} Data for the request
followers : {Array} An array of followers to add to the task.}
}
\description{
Adds each of the specified followers to the task, if they are not already
following. Returns the complete, updated record for the affected task.
}
