% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_add_subtask}
\alias{asn_tasks_add_subtask}
\title{Creates a new subtask and adds it to the parent task. Returns the full record
for the newly created subtask.}
\usage{
asn_tasks_add_subtask(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task to add a subtask to.#'}

\item{...}{{Object} Data for the request}
}
\description{
Creates a new subtask and adds it to the parent task. Returns the full record
for the newly created subtask.
}
