% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_add_tag}
\alias{asn_tasks_add_tag}
\title{Adds a tag to a task. Returns an empty data block.}
\usage{
asn_tasks_add_tag(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task to add a tag to.#'}

\item{...}{{Object} Data for the request
tag : {Id} The tag to add to the task.}
}
\description{
Adds a tag to a task. Returns an empty data block.
}
