% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_dependencies}
\alias{asn_tasks_dependencies}
\title{Returns the compact representations of all of the dependencies of a task.}
\usage{
asn_tasks_dependencies(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task to get dependencies on.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
Returns the compact representations of all of the dependencies of a task.
}
