% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_dependents}
\alias{asn_tasks_dependents}
\title{Returns the compact representations of all of the dependents of a task.}
\usage{
asn_tasks_dependents(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task to get dependents on.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
Returns the compact representations of all of the dependents of a task.
}
