% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_find_by_section}
\alias{asn_tasks_find_by_section}
\title{<b>Board view only:</b> Returns the compact section records for all tasks within the given section.}
\usage{
asn_tasks_find_by_section(section, ..., options = list())
}
\arguments{
\item{section}{{Id} The section in which to search for tasks.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
<b>Board view only:</b> Returns the compact section records for all tasks within the given section.
}
