% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_projects}
\alias{asn_tasks_projects}
\title{Returns a compact representation of all of the projects the task is in.}
\usage{
asn_tasks_projects(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task to get projects on.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
Returns a compact representation of all of the projects the task is in.
}
