% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_remove_dependents}
\alias{asn_tasks_remove_dependents}
\title{Unlinks a set of dependents from this task.}
\usage{
asn_tasks_remove_dependents(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task to remove dependents from.#'}

\item{...}{{Object} Data for the request
dependents : {Array} An array of task IDs to remove as dependents.}
}
\description{
Unlinks a set of dependents from this task.
}
