% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_remove_followers}
\alias{asn_tasks_remove_followers}
\title{Removes each of the specified followers from the task if they are
following. Returns the complete, updated record for the affected task.}
\usage{
asn_tasks_remove_followers(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task to remove followers from.#'}

\item{...}{{Object} Data for the request
followers : {Array} An array of followers to remove from the task.}
}
\description{
Removes each of the specified followers from the task if they are
following. Returns the complete, updated record for the affected task.
}
