% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_search}
\alias{asn_tasks_search}
\title{The search endpoint allows you to build complex queries to find and fetch exactly the data you need from Asana. For a more comprehensive description of all the query parameters and limitations of this endpoint, see our [long-form documentation](/developers/documentation/getting-started/search-api) for this feature.}
\usage{
asn_tasks_search(workspace, ..., options = list())
}
\arguments{
\item{workspace}{{Id} The workspace or organization in which to search for tasks.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
The search endpoint allows you to build complex queries to find and fetch exactly the data you need from Asana. For a more comprehensive description of all the query parameters and limitations of this endpoint, see our [long-form documentation](/developers/documentation/getting-started/search-api) for this feature.
}
