% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_stories}
\alias{asn_tasks_stories}
\title{Returns a compact representation of all of the stories on the task.}
\usage{
asn_tasks_stories(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task containing the stories to get.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
Returns a compact representation of all of the stories on the task.
}
