% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_subtasks}
\alias{asn_tasks_subtasks}
\title{Returns a compact representation of all of the subtasks of a task.}
\usage{
asn_tasks_subtasks(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task to get the subtasks of.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
Returns a compact representation of all of the subtasks of a task.
}
