% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{asn_teams}
\alias{asn_teams}
\title{Teams}
\description{
A _team_ is used to group related projects and people together within an
 organization. Each project in an organization is associated with a team.
}
\seealso{
\link{https://asana.com/developers/api-reference/teams}

Other teams: \code{\link{asn_teams_add_user}},
  \code{\link{asn_teams_find_by_id}},
  \code{\link{asn_teams_find_by_organization}},
  \code{\link{asn_teams_find_by_user}},
  \code{\link{asn_teams_remove_user}},
  \code{\link{asn_teams_users}}
}
\concept{teams}
