% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{asn_teams_add_user}
\alias{asn_teams_add_user}
\title{The user making this call must be a member of the team in order to add others.
The user to add must exist in the same organization as the team in order to be added.
The user to add can be referenced by their globally unique user ID or their email address.
Returns the full user record for the added user.}
\usage{
asn_teams_add_user(team, ..., options = list())
}
\arguments{
\item{team}{{Id} Globally unique identifier for the team.#'}

\item{...}{{Object} Data for the request
user : {String} An identifier for the user. Can be one of an email address,
the globally unique identifier for the user, or the keyword `me`
to indicate the current user making the request.}
}
\description{
The user making this call must be a member of the team in order to add others.
The user to add must exist in the same organization as the team in order to be added.
The user to add can be referenced by their globally unique user ID or their email address.
Returns the full user record for the added user.
}
\seealso{
Other teams: \code{\link{asn_teams_find_by_id}},
  \code{\link{asn_teams_find_by_organization}},
  \code{\link{asn_teams_find_by_user}},
  \code{\link{asn_teams_remove_user}},
  \code{\link{asn_teams_users}}, \code{\link{asn_teams}}
}
\concept{teams}
