% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{asn_teams_remove_user}
\alias{asn_teams_remove_user}
\title{The user to remove can be referenced by their globally unique user ID or their email address.
Removes the user from the specified team. Returns an empty data record.}
\usage{
asn_teams_remove_user(team, ..., options = list())
}
\arguments{
\item{team}{{Id} Globally unique identifier for the team.#'}

\item{...}{{Object} Data for the request
user : {String} An identifier for the user. Can be one of an email address,
the globally unique identifier for the user, or the keyword `me`
to indicate the current user making the request.}
}
\description{
The user to remove can be referenced by their globally unique user ID or their email address.
Removes the user from the specified team. Returns an empty data record.
}
