% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{asn_teams_users}
\alias{asn_teams_users}
\title{Returns the compact records for all users that are members of the team.}
\usage{
asn_teams_users(team, ..., options = list())
}
\arguments{
\item{team}{{Id} Globally unique identifier for the team.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
Returns the compact records for all users that are members of the team.
}
