% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{asn_users}
\alias{asn_users}
\title{Users}
\description{
A _user_ object represents an account in Asana that can be given access to
 various workspaces, projects, and tasks.
 
 Like other objects in the system, users are referred to by numerical IDs.
 However, the special string identifier `me` can be used anywhere
 a user ID is accepted, to refer to the current authenticated user.
}
\seealso{
\link{https://asana.com/developers/api-reference/users}

Other users: \code{\link{asn_users_find_all}},
  \code{\link{asn_users_find_by_id}},
  \code{\link{asn_users_find_by_workspace}},
  \code{\link{asn_users_me}}
}
