% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{asn_users_find_by_id}
\alias{asn_users_find_by_id}
\title{Returns the full user record for the single user with the provided ID.}
\usage{
asn_users_find_by_id(user, ..., options = list())
}
\arguments{
\item{user}{{String} An identifier for the user. Can be one of an email address,}

\item{...}{{Object} Parameters for the request}

\item{the}{globally unique identifier for the user, or the keyword `me`}

\item{to}{indicate the current user making the request.#'}
}
\description{
Returns the full user record for the single user with the provided ID.
}
\seealso{
Other users: \code{\link{asn_users_find_all}},
  \code{\link{asn_users_find_by_workspace}},
  \code{\link{asn_users_me}}, \code{\link{asn_users}}
}
