% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{asn_users_me}
\alias{asn_users_me}
\title{Returns the full user record for the currently authenticated user.}
\usage{
asn_users_me(..., options = list())
}
\arguments{
\item{...}{{Object} Parameters for the request}
}
\description{
Returns the full user record for the currently authenticated user.
}
\seealso{
Other users: \code{\link{asn_users_find_all}},
  \code{\link{asn_users_find_by_id}},
  \code{\link{asn_users_find_by_workspace}},
  \code{\link{asn_users}}
}
\concept{users}
