% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhooks.R
\name{asn_webhooks_get_all}
\alias{asn_webhooks_get_all}
\title{Returns the compact representation of all webhooks your app has
registered for the authenticated user in the given workspace.}
\usage{
asn_webhooks_get_all(..., options = list())
}
\arguments{
\item{...}{{Object} Parameters for the request
[resource] : {Id} Only return webhooks for the given resource.}

\item{workspace}{{Id} The workspace to query for webhooks in.#'}
}
\description{
Returns the compact representation of all webhooks your app has
registered for the authenticated user in the given workspace.
}
