% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{asn_workspaces}
\alias{asn_workspaces}
\title{Workspaces}
\description{
A _workspace_ is the highest-level organizational unit in Asana. All projects
 and tasks have an associated workspace.
 
 An _organization_ is a special kind of workspace that represents a company.
 In an organization, you can group your projects into teams. You can read
 more about how organizations work on the Asana Guide.
 To tell if your workspace is an organization or not, check its
 `is_organization` property.
 
 Over time, we intend to migrate most workspaces into organizations and to
 release more organization-specific functionality. We may eventually deprecate
 using workspace-based APIs for organizations. Currently, and until after
 some reasonable grace period following any further announcements, you can
 still reference organizations in any `workspace` parameter.
}
\seealso{
\link{https://asana.com/developers/api-reference/workspaces}

Other workspaces: \code{\link{asn_workspaces_add_user}},
  \code{\link{asn_workspaces_find_all}},
  \code{\link{asn_workspaces_find_by_id}},
  \code{\link{asn_workspaces_remove_user}},
  \code{\link{asn_workspaces_typeahead}},
  \code{\link{asn_workspaces_update}}
}
\concept{workspaces}
