% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{asn_workspaces_add_user}
\alias{asn_workspaces_add_user}
\title{The user can be referenced by their globally unique user ID or their email address.
Returns the full user record for the invited user.}
\usage{
asn_workspaces_add_user(workspace, ..., options = list())
}
\arguments{
\item{workspace}{{Id} The workspace or organization to invite the user to.#'}

\item{...}{{Object} Data for the request
user : {String} An identifier for the user. Can be one of an email address,
the globally unique identifier for the user, or the keyword `me`
to indicate the current user making the request.}
}
\description{
The user can be referenced by their globally unique user ID or their email address.
Returns the full user record for the invited user.
}
