% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{asn_workspaces_find_all}
\alias{asn_workspaces_find_all}
\title{Returns the compact records for all workspaces visible to the authorized user.}
\usage{
asn_workspaces_find_all(..., options = list())
}
\arguments{
\item{...}{{Object} Parameters for the request}
}
\description{
Returns the compact records for all workspaces visible to the authorized user.
}
\seealso{
Other workspaces: \code{\link{asn_workspaces_add_user}},
  \code{\link{asn_workspaces_find_by_id}},
  \code{\link{asn_workspaces_remove_user}},
  \code{\link{asn_workspaces_typeahead}},
  \code{\link{asn_workspaces_update}},
  \code{\link{asn_workspaces}}
}
\concept{workspaces}
