% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{asn_workspaces_find_by_id}
\alias{asn_workspaces_find_by_id}
\title{Returns the full workspace record for a single workspace.}
\usage{
asn_workspaces_find_by_id(workspace, ..., options = list())
}
\arguments{
\item{workspace}{{Id} Globally unique identifier for the workspace or organization.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
Returns the full workspace record for a single workspace.
}
