% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{asn_workspaces_typeahead}
\alias{asn_workspaces_typeahead}
\title{Retrieves objects in the workspace based on an auto-completion/typeahead
search algorithm. This feature is meant to provide results quickly, so do
not rely on this API to provide extremely accurate search results. The
result set is limited to a single page of results with a maximum size,
so you won't be able to fetch large numbers of results.}
\usage{
asn_workspaces_typeahead(workspace, ..., options = list())
}
\arguments{
\item{workspace}{{Id} The workspace to fetch objects from.#'}

\item{...}{{Object} Parameters for the request
type : {Enum} The type of values the typeahead should return. You can choose from
one of the following: custom_field, project, tag, task, and user.
Note that unlike in the names of endpoints, the types listed here are
in singular form (e.g. `task`). Using multiple types is not yet supported.
[query] : {String} The string that will be used to search for relevant objects. If an
empty string is passed in, the API will currently return an empty
result set.
[count] : {Number} The number of results to return. The default is `20` if this
parameter is omitted, with a minimum of `1` and a maximum of `100`.
If there are fewer results found than requested, all will be returned.}
}
\description{
Retrieves objects in the workspace based on an auto-completion/typeahead
search algorithm. This feature is meant to provide results quickly, so do
not rely on this API to provide extremely accurate search results. The
result set is limited to a single page of results with a maximum size,
so you won't be able to fetch large numbers of results.
}
\seealso{
Other workspaces: \code{\link{asn_workspaces_add_user}},
  \code{\link{asn_workspaces_find_all}},
  \code{\link{asn_workspaces_find_by_id}},
  \code{\link{asn_workspaces_remove_user}},
  \code{\link{asn_workspaces_update}},
  \code{\link{asn_workspaces}}
}
