% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{assert_is_access_token}
\alias{assert_is_access_token}
\title{Throw an error if the access token is malformed.}
\usage{
assert_is_access_token(access_token)
}
\description{
Throw an error if the access token is malformed.
}
