% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Set Asana Personal Access Token}
\usage{
authenticate(pat = NULL)
}
\arguments{
\item{pat}{Personal Access Token.}
}
\description{
Set Asana Personal Access Token
}
\examples{
\dontrun{
authenticate("1/889385868389540:a2af991203342d0be4fcde70e5d7c3cd")
}
}
