% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-for-projects.R
\name{get_all_projects}
\alias{get_all_projects}
\title{List all available projects in Asana}
\usage{
get_all_projects(archived = NA,
  access_token = Sys.getenv("ASANA_ACCESS_TOKEN"))
}
\arguments{
\item{archived}{TRUE, FALSE, or NA for archived, live, or all projects, respectively. Defaults to NA.}

\item{access_token}{An Asana access token.}
}
\value{
A list containing a dataframe of all available projects to the Asana user.
}
\description{
List all available projects in Asana
}
\examples{
\donttest{
## Marked as don't test because an access token is needed
get_all_projects()
}
}
\references{
\url{https://asana.com/developers/api-reference/projects#query}
}
