% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-asana-workspaces.R
\name{get_asana_workspaces}
\alias{get_asana_workspaces}
\title{List Available Workspaces in the Asana}
\usage{
get_asana_workspaces(access_token = Sys.getenv("ASANA_ACCESS_TOKEN"))
}
\arguments{
\item{access_token}{An Asana access token.}
}
\value{
A list containing a dataframe of all available workspaces to the Asana user.
}
\description{
List Available Workspaces in the Asana
}
\examples{
\donttest{
## Marked as don't test because an access token is needed
get_asana_workspaces()
}
}
\references{
\url{https://asana.com/developers/api-reference/workspaces#get}
}
