% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-for-projects.R
\name{get_workspace_projects}
\alias{get_workspace_projects}
\title{List a specific workspace's projects in Asana}
\usage{
get_workspace_projects(workspace_id, archived = FALSE,
  access_token = Sys.getenv("ASANA_ACCESS_TOKEN"))
}
\arguments{
\item{workspace_id}{An Asana workspace ID}

\item{archived}{A logical TRUE or FALSE to pull archived or live projects. Defaults to false.}

\item{access_token}{An Asana access token.}
}
\value{
A list containing a dataframe of all projects available in a specific workspace
}
\description{
List a specific workspace's projects in Asana
}
\examples{
\donttest{
## Marked as don't test because an access token is needed
get_workspace_projects()
}
}
\references{
\url{https://asana.com/developers/api-reference/projects#query}
}
