% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-asana-workspaces.R
\name{list_asana_workspaces}
\alias{list_asana_workspaces}
\title{List Available Workspaces in the Asana}
\usage{
list_asana_workspaces(access_token)
}
\value{
A list containing a dataframe of all available workspaces to the user
}
\description{
List Available Workspaces in the Asana
}
