% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{store_access_token}
\alias{store_access_token}
\title{Store an access token}
\usage{
store_access_token(access_token)
}
\arguments{
\item{access_token}{A string giving an access token to authenticate the API
call.}
}
\value{
Invoked for the side effect of setting the environment variable
\code{ASANA_ACCESS_TOKEN} to the input.
}
\description{
Stores an Asana API access token for the R session.
}
\details{
See the developer documentation for how to get an access token.
\url{https://asana.com/developers/documentation/getting-started/auth}
}
