% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users}
\alias{users}
\alias{me}
\alias{users_workspace}
\alias{users_gid}
\title{Get User Details}
\usage{
me()

users_workspace(workspace)

users_gid(gid)
}
\arguments{
\item{workspace}{Workspace ID.}

\item{gid}{User ID.}
}
\value{
A tibble.
}
\description{
Retrieve details of users.
}
\details{
The following information is available:
\itemize{
\item email
\item name
\item photos (various dimensions) and
\item workspaces.
}
}
\examples{
# Get details of authenticated user.
me()

# Get details of users in workspace.
users_workspace(1125279076968849)

# Get details of user by GID.
users_gid(887375868389530)
}
