% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fields.R
\name{asn_custom_fields_delete}
\alias{asn_custom_fields_delete}
\title{A specific, existing custom field can be deleted by making a DELETE request on the URL for that custom field.}
\usage{
asn_custom_fields_delete(custom_field, ..., options = list())
}
\arguments{
\item{custom_field}{{Id} Globally unique identifier for the custom field.}
}
\description{
Returns an empty data record.
}
\seealso{
Other custom_fields: \code{\link{asn_custom_fields_add_enum_option}},
  \code{\link{asn_custom_fields_create}},
  \code{\link{asn_custom_fields_find_by_id}},
  \code{\link{asn_custom_fields_find_by_workspace}},
  \code{\link{asn_custom_fields_reorder_enum_option}},
  \code{\link{asn_custom_fields_update_enum_option}},
  \code{\link{asn_custom_fields_update}},
  \code{\link{asn_custom_fields}}
}
