% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_projects_add_members}
\alias{asn_projects_add_members}
\title{Adds the specified list of users as members of the project. Returns the updated project record.}
\usage{
asn_projects_add_members(project, ..., options = list())
}
\arguments{
\item{project}{{Id} The project to add members to.#'}

\item{...}{{Object} Data for the request
members : {Array} An array of members to add to the project.}
}
\description{
Adds the specified list of users as members of the project. Returns the updated project record.
}
\seealso{
Other projects: \code{\link{asn_projects_add_custom_field_setting}},
  \code{\link{asn_projects_add_followers}},
  \code{\link{asn_projects_create_in_team}},
  \code{\link{asn_projects_create_in_workspace}},
  \code{\link{asn_projects_create}},
  \code{\link{asn_projects_delete}},
  \code{\link{asn_projects_find_all}},
  \code{\link{asn_projects_find_by_id}},
  \code{\link{asn_projects_find_by_team}},
  \code{\link{asn_projects_find_by_workspace}},
  \code{\link{asn_projects_remove_custom_field_setting}},
  \code{\link{asn_projects_remove_followers}},
  \code{\link{asn_projects_remove_members}},
  \code{\link{asn_projects_tasks}},
  \code{\link{asn_projects_update}},
  \code{\link{asn_projects}}
}
\concept{projects}
