% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_projects_delete}
\alias{asn_projects_delete}
\title{A specific, existing project can be deleted by making a DELETE request
on the URL for that project.}
\usage{
asn_projects_delete(project, ..., options = list())
}
\arguments{
\item{project}{{Id} The project to delete.}
}
\description{
Returns an empty data record.
}
