% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{asn_projects_find_by_team}
\alias{asn_projects_find_by_team}
\title{Returns the compact project records for all projects in the team.}
\usage{
asn_projects_find_by_team(team, ..., options = list())
}
\arguments{
\item{team}{{Id} The team to find projects in.#'}

\item{...}{{Object} Parameters for the request
[archived] : {Boolean} Only return projects whose `archived` field takes on the value of
this parameter.}
}
\description{
Returns the compact project records for all projects in the team.
}
\seealso{
Other projects: \code{\link{asn_projects_add_custom_field_setting}},
  \code{\link{asn_projects_add_followers}},
  \code{\link{asn_projects_add_members}},
  \code{\link{asn_projects_create_in_team}},
  \code{\link{asn_projects_create_in_workspace}},
  \code{\link{asn_projects_create}},
  \code{\link{asn_projects_delete}},
  \code{\link{asn_projects_find_all}},
  \code{\link{asn_projects_find_by_id}},
  \code{\link{asn_projects_find_by_workspace}},
  \code{\link{asn_projects_remove_custom_field_setting}},
  \code{\link{asn_projects_remove_followers}},
  \code{\link{asn_projects_remove_members}},
  \code{\link{asn_projects_tasks}},
  \code{\link{asn_projects_update}},
  \code{\link{asn_projects}}
}
