% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{asn_sections_find_by_id}
\alias{asn_sections_find_by_id}
\title{Returns the complete record for a single section.}
\usage{
asn_sections_find_by_id(section, ..., options = list())
}
\arguments{
\item{section}{{Id} The section to get.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
Returns the complete record for a single section.
}
\seealso{
Other sections: \code{\link{asn_sections_create_in_project}},
  \code{\link{asn_sections_delete}},
  \code{\link{asn_sections_find_by_project}},
  \code{\link{asn_sections_insert_in_project}},
  \code{\link{asn_sections_update}},
  \code{\link{asn_sections}}
}
\concept{sections}
