% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{asn_stories_delete}
\alias{asn_stories_delete}
\title{Deletes a story. A user can only delete stories they have created. Returns an empty data record.}
\usage{
asn_stories_delete(story, ..., options = list())
}
\arguments{
\item{story}{{Id} Globally unique identifier for the story.}
}
\description{
Deletes a story. A user can only delete stories they have created. Returns an empty data record.
}
\seealso{
Other stories: \code{\link{asn_stories_create_on_task}},
  \code{\link{asn_stories_find_by_id}},
  \code{\link{asn_stories_find_by_task}},
  \code{\link{asn_stories_update}},
  \code{\link{asn_stories}}
}
\concept{stories}
