% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{asn_stories_find_by_task}
\alias{asn_stories_find_by_task}
\title{Returns the compact records for all stories on the task.}
\usage{
asn_stories_find_by_task(task, ..., options = list())
}
\arguments{
\item{task}{{Id} Globally unique identifier for the task.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
Returns the compact records for all stories on the task.
}
\seealso{
Other stories: \code{\link{asn_stories_create_on_task}},
  \code{\link{asn_stories_delete}},
  \code{\link{asn_stories_find_by_id}},
  \code{\link{asn_stories_update}},
  \code{\link{asn_stories}}
}
