% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{asn_tags_delete}
\alias{asn_tags_delete}
\title{A specific, existing tag can be deleted by making a DELETE request
on the URL for that tag.}
\usage{
asn_tags_delete(tag, ..., options = list())
}
\arguments{
\item{tag}{{Id} The tag to delete.}
}
\description{
Returns an empty data record.
}
\seealso{
Other tags: \code{\link{asn_tags_create_in_workspace}},
  \code{\link{asn_tags_create}},
  \code{\link{asn_tags_find_all}},
  \code{\link{asn_tags_find_by_id}},
  \code{\link{asn_tags_find_by_workspace}},
  \code{\link{asn_tags_get_tasks_with_tag}},
  \code{\link{asn_tags_update}}, \code{\link{asn_tags}}
}
