% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_add_comment}
\alias{asn_tasks_add_comment}
\title{Adds a comment to a task. The comment will be authored by the
currently authenticated user, and timestamped when the server receives
the request.}
\usage{
asn_tasks_add_comment(task, ..., options = list())
}
\arguments{
\item{task}{{Id} Globally unique identifier for the task.#'}

\item{...}{{Object} Data for the request
text : {String} The plain text of the comment to add.}
}
\description{
Returns the full record for the new story added to the task.
}
\seealso{
Other tasks: \code{\link{asn_tasks_add_dependencies}},
  \code{\link{asn_tasks_add_dependents}},
  \code{\link{asn_tasks_add_followers}},
  \code{\link{asn_tasks_add_project}},
  \code{\link{asn_tasks_add_subtask}},
  \code{\link{asn_tasks_add_tag}},
  \code{\link{asn_tasks_create_in_workspace}},
  \code{\link{asn_tasks_create}},
  \code{\link{asn_tasks_delete}},
  \code{\link{asn_tasks_dependencies}},
  \code{\link{asn_tasks_dependents}},
  \code{\link{asn_tasks_find_all}},
  \code{\link{asn_tasks_find_by_id}},
  \code{\link{asn_tasks_find_by_project}},
  \code{\link{asn_tasks_find_by_section}},
  \code{\link{asn_tasks_find_by_tag}},
  \code{\link{asn_tasks_projects}},
  \code{\link{asn_tasks_remove_dependencies}},
  \code{\link{asn_tasks_remove_dependents}},
  \code{\link{asn_tasks_remove_followers}},
  \code{\link{asn_tasks_remove_project}},
  \code{\link{asn_tasks_remove_tag}},
  \code{\link{asn_tasks_search}},
  \code{\link{asn_tasks_stories}},
  \code{\link{asn_tasks_subtasks}},
  \code{\link{asn_tasks_tags}},
  \code{\link{asn_tasks_update}}, \code{\link{asn_tasks}}
}
