% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_add_dependencies}
\alias{asn_tasks_add_dependencies}
\title{Marks a set of tasks as dependencies of this task, if they are not
already dependencies. *A task can have at most 15 dependencies.*}
\usage{
asn_tasks_add_dependencies(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task to add dependencies to.#'}

\item{...}{{Object} Data for the request
dependencies : {Array} An array of task IDs that this task should depend on.}
}
\description{
Marks a set of tasks as dependencies of this task, if they are not
already dependencies. *A task can have at most 15 dependencies.*
}
\seealso{
Other tasks: \code{\link{asn_tasks_add_comment}},
  \code{\link{asn_tasks_add_dependents}},
  \code{\link{asn_tasks_add_followers}},
  \code{\link{asn_tasks_add_project}},
  \code{\link{asn_tasks_add_subtask}},
  \code{\link{asn_tasks_add_tag}},
  \code{\link{asn_tasks_create_in_workspace}},
  \code{\link{asn_tasks_create}},
  \code{\link{asn_tasks_delete}},
  \code{\link{asn_tasks_dependencies}},
  \code{\link{asn_tasks_dependents}},
  \code{\link{asn_tasks_find_all}},
  \code{\link{asn_tasks_find_by_id}},
  \code{\link{asn_tasks_find_by_project}},
  \code{\link{asn_tasks_find_by_section}},
  \code{\link{asn_tasks_find_by_tag}},
  \code{\link{asn_tasks_projects}},
  \code{\link{asn_tasks_remove_dependencies}},
  \code{\link{asn_tasks_remove_dependents}},
  \code{\link{asn_tasks_remove_followers}},
  \code{\link{asn_tasks_remove_project}},
  \code{\link{asn_tasks_remove_tag}},
  \code{\link{asn_tasks_search}},
  \code{\link{asn_tasks_stories}},
  \code{\link{asn_tasks_subtasks}},
  \code{\link{asn_tasks_tags}},
  \code{\link{asn_tasks_update}}, \code{\link{asn_tasks}}
}
\concept{tasks}
