% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_add_subtask}
\alias{asn_tasks_add_subtask}
\title{Creates a new subtask and adds it to the parent task. Returns the full record
for the newly created subtask.}
\usage{
asn_tasks_add_subtask(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task to add a subtask to.#'}

\item{...}{{Object} Data for the request}
}
\description{
Creates a new subtask and adds it to the parent task. Returns the full record
for the newly created subtask.
}
\seealso{
Other tasks: \code{\link{asn_tasks_add_comment}},
  \code{\link{asn_tasks_add_dependencies}},
  \code{\link{asn_tasks_add_dependents}},
  \code{\link{asn_tasks_add_followers}},
  \code{\link{asn_tasks_add_project}},
  \code{\link{asn_tasks_add_tag}},
  \code{\link{asn_tasks_create_in_workspace}},
  \code{\link{asn_tasks_create}},
  \code{\link{asn_tasks_delete}},
  \code{\link{asn_tasks_dependencies}},
  \code{\link{asn_tasks_dependents}},
  \code{\link{asn_tasks_find_all}},
  \code{\link{asn_tasks_find_by_id}},
  \code{\link{asn_tasks_find_by_project}},
  \code{\link{asn_tasks_find_by_section}},
  \code{\link{asn_tasks_find_by_tag}},
  \code{\link{asn_tasks_projects}},
  \code{\link{asn_tasks_remove_dependencies}},
  \code{\link{asn_tasks_remove_dependents}},
  \code{\link{asn_tasks_remove_followers}},
  \code{\link{asn_tasks_remove_project}},
  \code{\link{asn_tasks_remove_tag}},
  \code{\link{asn_tasks_search}},
  \code{\link{asn_tasks_stories}},
  \code{\link{asn_tasks_subtasks}},
  \code{\link{asn_tasks_tags}},
  \code{\link{asn_tasks_update}}, \code{\link{asn_tasks}}
}
\concept{tasks}
