% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_create}
\alias{asn_tasks_create}
\title{Creating a new task is as easy as POSTing to the `/tasks` endpoint
with a data block containing the fields you'd like to set on the task.
Any unspecified fields will take on default values.}
\usage{
asn_tasks_create(..., options = list())
}
\arguments{
\item{...}{{Object} Data for the request
[workspace] : {Id} The workspace to create a task in.}
}
\description{
Every task is required to be created in a specific workspace, and this
workspace cannot be changed once set. The workspace need not be set
explicitly if you specify `projects` or a `parent` task instead.
}
\details{
`projects` can be a comma separated list of projects, or just a single
project the task should belong to.
}
