% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_find_all}
\alias{asn_tasks_find_all}
\title{Returns the compact task records for some filtered set of tasks. Use one
or more of the parameters provided to filter the tasks returned. You must
specify a `project` or `tag` if you do not specify `assignee` and `workspace`.}
\usage{
asn_tasks_find_all(..., options = list())
}
\arguments{
\item{...}{{Object} Parameters for the request
[assignee] : {String} The assignee to filter tasks on.
[project] : {Id} The project to filter tasks on.
[section] : {Id} The section to filter tasks on.
[workspace] : {Id} The workspace or organization to filter tasks on.
[completed_since] : {String} Only return tasks that are either incomplete or that have been
completed since this time.
[modified_since] : {String} Only return tasks that have been modified since the given time.}
}
\description{
Returns the compact task records for some filtered set of tasks. Use one
or more of the parameters provided to filter the tasks returned. You must
specify a `project` or `tag` if you do not specify `assignee` and `workspace`.
}
\seealso{
Other tasks: \code{\link{asn_tasks_add_comment}},
  \code{\link{asn_tasks_add_dependencies}},
  \code{\link{asn_tasks_add_dependents}},
  \code{\link{asn_tasks_add_followers}},
  \code{\link{asn_tasks_add_project}},
  \code{\link{asn_tasks_add_subtask}},
  \code{\link{asn_tasks_add_tag}},
  \code{\link{asn_tasks_create_in_workspace}},
  \code{\link{asn_tasks_create}},
  \code{\link{asn_tasks_delete}},
  \code{\link{asn_tasks_dependencies}},
  \code{\link{asn_tasks_dependents}},
  \code{\link{asn_tasks_find_by_id}},
  \code{\link{asn_tasks_find_by_project}},
  \code{\link{asn_tasks_find_by_section}},
  \code{\link{asn_tasks_find_by_tag}},
  \code{\link{asn_tasks_projects}},
  \code{\link{asn_tasks_remove_dependencies}},
  \code{\link{asn_tasks_remove_dependents}},
  \code{\link{asn_tasks_remove_followers}},
  \code{\link{asn_tasks_remove_project}},
  \code{\link{asn_tasks_remove_tag}},
  \code{\link{asn_tasks_search}},
  \code{\link{asn_tasks_stories}},
  \code{\link{asn_tasks_subtasks}},
  \code{\link{asn_tasks_tags}},
  \code{\link{asn_tasks_update}}, \code{\link{asn_tasks}}
}
\concept{tasks}
