% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_remove_followers}
\alias{asn_tasks_remove_followers}
\title{Removes each of the specified followers from the task if they are
following. Returns the complete, updated record for the affected task.}
\usage{
asn_tasks_remove_followers(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task to remove followers from.#'}

\item{...}{{Object} Data for the request
followers : {Array} An array of followers to remove from the task.}
}
\description{
Removes each of the specified followers from the task if they are
following. Returns the complete, updated record for the affected task.
}
\seealso{
Other tasks: \code{\link{asn_tasks_add_comment}},
  \code{\link{asn_tasks_add_dependencies}},
  \code{\link{asn_tasks_add_dependents}},
  \code{\link{asn_tasks_add_followers}},
  \code{\link{asn_tasks_add_project}},
  \code{\link{asn_tasks_add_subtask}},
  \code{\link{asn_tasks_add_tag}},
  \code{\link{asn_tasks_create_in_workspace}},
  \code{\link{asn_tasks_create}},
  \code{\link{asn_tasks_delete}},
  \code{\link{asn_tasks_dependencies}},
  \code{\link{asn_tasks_dependents}},
  \code{\link{asn_tasks_find_all}},
  \code{\link{asn_tasks_find_by_id}},
  \code{\link{asn_tasks_find_by_project}},
  \code{\link{asn_tasks_find_by_section}},
  \code{\link{asn_tasks_find_by_tag}},
  \code{\link{asn_tasks_projects}},
  \code{\link{asn_tasks_remove_dependencies}},
  \code{\link{asn_tasks_remove_dependents}},
  \code{\link{asn_tasks_remove_project}},
  \code{\link{asn_tasks_remove_tag}},
  \code{\link{asn_tasks_search}},
  \code{\link{asn_tasks_stories}},
  \code{\link{asn_tasks_subtasks}},
  \code{\link{asn_tasks_tags}},
  \code{\link{asn_tasks_update}}, \code{\link{asn_tasks}}
}
