% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_tags}
\alias{asn_tasks_tags}
\title{Returns a compact representation of all of the tags the task has.}
\usage{
asn_tasks_tags(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task to get tags on.#'}

\item{...}{{Object} Parameters for the request}
}
\description{
Returns a compact representation of all of the tags the task has.
}
