% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{asn_tasks_update}
\alias{asn_tasks_update}
\title{A specific, existing task can be updated by making a PUT request on the
URL for that task. Only the fields provided in the `data` block will be
updated; any unspecified fields will remain unchanged.}
\usage{
asn_tasks_update(task, ..., options = list())
}
\arguments{
\item{task}{{Id} The task to update.#'}

\item{...}{{Object} Data for the request}
}
\description{
When using this method, it is best to specify only those fields you wish
to change, or else you may overwrite changes made by another user since
you last retrieved the task.
}
\details{
Returns the complete updated task record.
}
\seealso{
Other tasks: \code{\link{asn_tasks_add_comment}},
  \code{\link{asn_tasks_add_dependencies}},
  \code{\link{asn_tasks_add_dependents}},
  \code{\link{asn_tasks_add_followers}},
  \code{\link{asn_tasks_add_project}},
  \code{\link{asn_tasks_add_subtask}},
  \code{\link{asn_tasks_add_tag}},
  \code{\link{asn_tasks_create_in_workspace}},
  \code{\link{asn_tasks_create}},
  \code{\link{asn_tasks_delete}},
  \code{\link{asn_tasks_dependencies}},
  \code{\link{asn_tasks_dependents}},
  \code{\link{asn_tasks_find_all}},
  \code{\link{asn_tasks_find_by_id}},
  \code{\link{asn_tasks_find_by_project}},
  \code{\link{asn_tasks_find_by_section}},
  \code{\link{asn_tasks_find_by_tag}},
  \code{\link{asn_tasks_projects}},
  \code{\link{asn_tasks_remove_dependencies}},
  \code{\link{asn_tasks_remove_dependents}},
  \code{\link{asn_tasks_remove_followers}},
  \code{\link{asn_tasks_remove_project}},
  \code{\link{asn_tasks_remove_tag}},
  \code{\link{asn_tasks_search}},
  \code{\link{asn_tasks_stories}},
  \code{\link{asn_tasks_subtasks}},
  \code{\link{asn_tasks_tags}}, \code{\link{asn_tasks}}
}
\concept{tasks}
