% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{asn_workspaces_find_all}
\alias{asn_workspaces_find_all}
\title{Returns the compact records for all workspaces visible to the authorized user.}
\usage{
asn_workspaces_find_all(..., options = list())
}
\arguments{
\item{...}{{Object} Parameters for the request}
}
\description{
Returns the compact records for all workspaces visible to the authorized user.
}
