% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{asn_workspaces_update}
\alias{asn_workspaces_update}
\title{A specific, existing workspace can be updated by making a PUT request on
the URL for that workspace. Only the fields provided in the data block
will be updated; any unspecified fields will remain unchanged.}
\usage{
asn_workspaces_update(workspace, ..., options = list())
}
\arguments{
\item{workspace}{{Id} The workspace to update.#'}

\item{...}{{Object} Data for the request}
}
\description{
Currently the only field that can be modified for a workspace is its `name`.
}
\details{
Returns the complete, updated workspace record.
}
\seealso{
Other workspaces: \code{\link{asn_workspaces_add_user}},
  \code{\link{asn_workspaces_find_all}},
  \code{\link{asn_workspaces_find_by_id}},
  \code{\link{asn_workspaces_remove_user}},
  \code{\link{asn_workspaces_typeahead}},
  \code{\link{asn_workspaces}}
}
\concept{workspaces}
