% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gid}
\alias{get_gid}
\alias{get_gid.asana_api}
\alias{get_gid.data.frame}
\title{Get the Global ID}
\usage{
get_gid(x)

\method{get_gid}{asana_api}(x)

\method{get_gid}{data.frame}(x)
}
\arguments{
\item{x}{An Asana object.}
}
\value{
The global ID, as a character vector of length 1.
}
\description{
Gets the global ID for an Asana object.
}
